change the dropped rate message to focus on unique email addresses
Showing that 20/20 emails to a single address failed isn't all that
useful. What we really want is to be alerted when the majority of
addresses at a single domain are rejecting our email.

This change means we'll only be alerted when > 50% of unique email
addresses at a domain are failing. In practice, that means we need at
least 3 unique addresses to have received email from us before an alert
is possible.